Consolidate multiple mock_chat_completion into one
