+++ test/components/generators/test_openai.py
@@ -1,42 +1,12 @@
 import os
-from datetime import datetime
 from typing import List
-from unittest.mock import patch, Mock
 
 import pytest
 from openai import OpenAIError
-from openai.types.chat import ChatCompletion, ChatCompletionMessage
-from openai.types.chat.chat_completion import Choice
 
 from haystack.components.generators import GPTGenerator
 from haystack.components.generators.utils import default_streaming_callback
 from haystack.dataclasses import StreamingChunk, ChatMessage
-
-
-@pytest.fixture
-def mock_chat_completion():
-    """
-    Mock the OpenAI API completion response and reuse it for tests
-    """
-    with patch("openai.resources.chat.completions.Completions.create") as mock_chat_completion_create:
-        completion = ChatCompletion(
-            id="foo",
-            model="gpt-4",
-            object="chat.completion",
-            choices=[
-                Choice(
-                    finish_reason="stop",
-                    logprobs=None,
-                    index=0,
-                    message=ChatCompletionMessage(content="Hello world!", role="assistant"),
-                )
-            ],
-            created=int(datetime.now().timestamp()),
-            usage={"prompt_tokens": 57, "completion_tokens": 40, "total_tokens": 97},
-        )
-
-        mock_chat_completion_create.return_value = completion
-        yield mock_chat_completion_create
 
 
 class TestGPTGenerator:
